requirements.txt
boto3-stubs[essential,polly,mediaconvert]==1.34.*
boto3==1.34.*
celery-types==0.14.0
celery==5.2.7
dj-database-url==1.0.0
django_jsonform==2.15.*
django-loginas==0.3.11
django-stubs-ext==0.5.0
django-stubs==1.12.0
Django==5.0.3
djangorestframework-dataclasses==1.2.*
djangorestframework-stubs==1.7.0
djangorestframework==3.14.0
drf-yasg-stubs==0.1.*
drf-yasg==1.21.*
factory_boy==3.2.1
Faker==15.0.0
fakeredis==2.16.*
filelock==3.*
flake8==5.0.4
gunicorn==20.1.0
mypy-extensions==0.4.3
mypy==0.961
psycopg2-binary==2.9.3
pytest-cov==4.0.0
pytest-django==4.5.2
pytest==7.1.3
python-dotenv==0.21.*
python-magic==0.4.27
redis==4.3.4
ruff==0.1.13
sentry-sdk==1.40.4
types-dj-database-url==1.0.0
whitenoise==6.2.0
